[DPE-7846] fix: secrets not set issue in RelationState.update
